Move string functions to the separate files.
